+++ src/config.ts
@@ -272,7 +272,6 @@ function getConfig() {
     fiddleFilePaths: configOptions.get<fiddleFilesUtil.FiddleFilePaths>('fiddleFilePaths'),
     outputDestinations:
       configOptions.get<output.OutputDestinationConfiguration>('outputDestinations'),
-    useLegacyReplWindowPath: configOptions.get<boolean>('useLegacyReplWindowPath'),
     legacyPrintBareReplWindowOutput: configOptions.get<boolean>('legacyPrintBareReplWindowOutput'),
     basilispPath: configOptions.get<string>('basilispPath'),
     refreshNssBeforeFn: configOptions.get<string>('refreshNssBeforeFn'),