+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Changes to Calva.
   - Adds a REPL Sessions status bar menu for pinning sessions, toggling auto-route, and setting `.cljc` overrides.
   - Adds API function `repl.listSessions()`
   - Updates default connect sequences for *Joyride*, *Babashka*, *nbb*, and *Basilisp*, to use custom `replSessionNames`, and configured default routings with `replSessionGlobs`.
+  - Removed deprecated `calva.useLegacyReplWindowPath` setting. The REPL window is now always located at `.calva/repl.calva-repl`.
 
 ## [2.0.540] - 2025-11-06